+++ src/sections/Projects/Sistent/components/select/guidance.js
@@ -13,10 +13,8 @@ import {
   OutlinedInput,
 } from "@mui/material";
 import { useStyledDarkMode } from "../../../../../theme/app/useStyledDarkMode";
-import { useLocation } from "@reach/router";
-import { navigate } from "gatsby";
-import TabButton from "../../../../../reusecore/Button";
-
+import SectionNav from "./section-nav";
+import Header from "./header";
 const SelectGuidance = () => {
   const { isDark } = useStyledDarkMode();
   const [selectedAge, setSelectedAge] = React.useState("");
@@ -25,45 +23,11 @@ const SelectGuidance = () => {
     let agesList = event.target.value;
     setMultipleAges(agesList);
   };
-  const location = useLocation();
   return (
     <SistentLayout title="Select">
       <div className="content">
-        <a id="Identity">
-          <h2>Select</h2>
-        </a>
-        <p>
-        Select component is a dropdown menu for selecting an option from a list.
-        </p>
-        <div className="filterBtns">
-          <TabButton
-            className={
-              location.pathname === "/projects/sistent/components/select"
-                ? "active"
-                : ""
-            }
-            onClick={() => navigate("/projects/sistent/components/select")}
-            title="Overview"
-          />
-          <TabButton
-            className={
-              location.pathname === "/projects/sistent/components/select/guidance"
-                ? "active"
-                : ""
-            }
-            onClick={() => navigate("/projects/sistent/components/select/guidance")}
-            title="Guidance"
-          />
-          <TabButton
-            className={
-              location.pathname === "/projects/sistent/components/select/code"
-                ? "active"
-                : ""
-            }
-            onClick={() => navigate("/projects/sistent/components/select/code")}
-            title="Code"
-          />
-        </div>
+        <Header />
+        <SectionNav />
         <div className="main-content">
           <a id="Functions">
             <h2>Functions</h2>
